asana-rl-seed-data/requirements.txt
faker
python-dotenv
openai
pandas
tqdm